move escape and unescape from StringUtil to IdUtil
