+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/id/IdUtil.java
@@ -19,6 +19,8 @@
 
 package com.baidu.hugegraph.backend.id;
 
+import org.apache.commons.lang3.StringUtils;
+
 import com.baidu.hugegraph.util.E;
 
 public class IdUtil {
@@ -44,5 +46,47 @@ public class IdUtil {
         } else {
             return IdGenerator.of(id);
         }
+    }
+
+    public static String escape(char splitor, char escape, String... values) {
+        StringBuilder escaped = new StringBuilder((values.length + 1) << 4);
+        // Do escape for every item in values
+        for (String value : values) {
+            if (escaped.length() > 0) {
+                escaped.append(splitor);
+            }
+
+            if (value.indexOf(splitor) == -1) {
+                escaped.append(value);
+                continue;
+            }
+
+            // Do escape for current item
+            for (int i = 0, n = value.length(); i < n; i++) {
+                char ch = value.charAt(i);
+                if (ch == splitor) {
+                    escaped.append(escape);
+                }
+                escaped.append(ch);
+            }
+        }
+        return escaped.toString();
+    }
+
+    public static String[] unescape(String id, String splitor, String escape) {
+        /*
+         * Note that the `splitor`/`escape` maybe special characters in regular
+         * expressions, but this is a frequently called method, for faster
+         * execution, we forbid the use of special characters as delimiter
+         * or escape sign.
+         * The `limit` param -1 in split method can ensure empty string be
+         * splited to a part.
+         */
+        String[] parts = id.split("(?<!" + escape + ")" + splitor, -1);
+        for (int i = 0; i < parts.length; i++) {
+            parts[i] = StringUtils.replace(parts[i], escape + splitor,
+                                           splitor);
+        }
+        return parts;
     }
 }
